collect_data/requirements.txt
requests~=2.25.1
# pymongo~=3.11.3
google-cloud-firestore~=2.1.0